添加了gui的系统调用

+++ include/int.h
@@ -142,8 +142,8 @@ func_num,char *str
 12:
 func,count,char *dist
  */
-int   syscall(long a, long b, long c, long d, long e, long f);
-void* _syscall(int func_num, ...);
+unsigned long long syscall(long a, long b, long c, long d, long e, long f);
+void*              _syscall(int func_num, ...);
 
 
 #endif
